feat: add mine bottom sheet component
