+++ src/components/community/MineBottomSheet/styled.ts
@@ -1,0 +1,20 @@
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
+
+import styled from '@emotion/native';
+
+export const CommunityMineBottomSheetContainer = styled.View`
+  row-gap: 30px;
+  padding: 14px;
+  z-index: 9999;
+`;
+
+export const CommunityMineOptionContainer = styled.View`
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+`;
+
+export const CommunityMainOptionIconContainer = styled(CommunityMineOptionContainer)`
+  column-gap: 10px;
+  justify-content: center;
+`;
