+++ src/screens/community/mine/styled.ts
@@ -27,20 +27,3 @@ export const CommunityMinePostBox = styled.View`
   padding: 10px 0;
   position: relative;
 `;
-
-export const CommunityMineBottomSheetContainer = styled.View`
-  row-gap: 30px;
-  padding: 14px;
-  z-index: 9999;
-`;
-
-export const CommunityMineOptionContainer = styled.View`
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-export const CommunityMainOptionIconContainer = styled(CommunityMineOptionContainer)`
-  column-gap: 10px;
-  justify-content: center;
-`;
